refactor: move nodes into categorical modules
